+++ experimental/packages/opentelemetry-sdk-node/test/sdk.test.ts
@@ -1017,6 +1017,41 @@ describe('Node SDK', () => {
       });
       await sdk.shutdown();
     });
+
+    it('should configure service instance id with service instance id from env variable taking priority over random UUID, based on order', async () => {
+      process.env.OTEL_RESOURCE_ATTRIBUTES =
+        'service.instance.id=custom-service,service.name=my-service';
+      process.env.OTEL_NODE_RESOURCE_DETECTORS = 'serviceinstance,env';
+      const sdk = new NodeSDK({
+        autoDetectResources: true,
+      });
+
+      sdk.start();
+      const resource = sdk['_resource'];
+      await resource.waitForAsyncAttributes?.();
+
+      assertServiceResource(resource, {
+        name: 'my-service',
+        instanceId: 'custom-service',
+      });
+      await sdk.shutdown();
+    });
+
+    it('should configure service instance id with service instance id from env variable taking priority over random UUID, based on order', async () => {
+      process.env.OTEL_RESOURCE_ATTRIBUTES =
+        'service.instance.id=custom-service,service.name=my-service';
+      process.env.OTEL_NODE_RESOURCE_DETECTORS = 'env,serviceinstance';
+      const sdk = new NodeSDK({
+        autoDetectResources: true,
+      });
+
+      sdk.start();
+      const resource = sdk['_resource'];
+      await resource.waitForAsyncAttributes?.();
+
+      assertServiceInstanceIdIsUUID(resource);
+      await sdk.shutdown();
+    });
   });
 
   describe('A disabled SDK should be no-op', () => {